Feature parity!

+++ arm64/archdefs.h
@@ -2,10 +2,10 @@ typedef struct TrapFrame TrapFrame;
 
 // arm64/syscall.c
 void syscall(TrapFrame *tf);
-long argfd(int n, int *fd);
-long argint(int n, int *ip);
-long arglong(int n, long *lp);
-long argptr(int n, uintptr *p, usize size);
+int argfd(int n, int *fd);
+int argint(int n, int *ip);
+int arglong(int n, long *lp);
+int argptr(int n, uintptr *p, usize size);
 long argstr(int n, char **pp);
 
 // proc.c
